refactor update subtitle
